firehose-direct-put/requirements.txt
boto3>=1.34.110
python-dotenv>=1.0.1
